ci: dry-run release v0.14.2

+++ reproducible-builds/build.sh
@@ -18,7 +18,7 @@ fi
 VERSION=$(sed -n 's/\s*version\s*=\s*"\(.*\)".*/\1/p' build.gradle.kts | tail -n1)
 echo "$VERSION" >dist/VERSION
 
-$ENGINE build -t signal-cli:build -f reproducible-builds/build.Containerfile .
+$ENGINE build -t signal-cli:build ${OVERRIDE_JAVA_VERSION:+--build-arg ZULU_TAG=$OVERRIDE_JAVA_VERSION} -f reproducible-builds/build.Containerfile .
 $ENGINE build -t signal-cli:native -f reproducible-builds/native.Containerfile .
 $ENGINE build -t signal-cli:client -f reproducible-builds/client.Containerfile .
 
@@ -27,6 +27,11 @@ git clean -Xfd -e '!/dist/' -e '!/dist/**' -e '!/github/' -e '!/github/**'
 # shellcheck disable=SC2086
 $ENGINE run --pull=never --rm -v "$(pwd)":/signal-cli:Z -e VERSION="$VERSION" $USER signal-cli:build
 mv build/distributions/signal-cli-*.tar.gz dist/
+
+if [ -n "${OVERRIDE_JAVA_VERSION:-}" ]; then
+	echo -e "\e[33mBuild was performed with overridden Java version $OVERRIDE_JAVA_VERSION, native-image and client will not be built.\e[0m"
+	exit 0
+fi
 
 # Build native-image
 git clean -Xfd -e '!/dist/' -e '!/dist/**' -e '!/github/' -e '!/github/**'
